+++ iOSClient/Viewer/NCViewerRichdocument/NCViewerRichdocument.swift
@@ -181,7 +181,7 @@ class NCViewerRichdocument: UIViewController, WKNavigationDelegate, WKScriptMess
             }
 
             if message.body as? String == "share" {
-                NCFunctionCenter.shared.openShare(ViewController: self, metadata: metadata, indexPage: .sharing)
+                NCFunctionCenter.shared.openShare(viewController: self, metadata: metadata, indexPage: .sharing)
             }
 
             if let param = message.body as? [AnyHashable: Any] {
